bugfix: If multiple IFDs are present, seek to them before reading from the directory
== Background ==
TIFF files contain one ore more Image File Directories (IFDs). These IFD blocks are chained so that the offset to the next IFD block is stored at the end of the current IFD block, ending when the offset is zero. The offset to the first IFD block is stored in the TIFF header.

== Bug ==
When reading a TIFF file with multiple IFD blocks, we presently do not seek to the next IFD at the end of the presently processed IFD block. Thus, we end up reading garbled data.

== Fix ==
Seek to the offset of the IFD block at the beginning of processing each IFD, not just for the first IFD.